+++ files/index.d.ts
@@ -1063,12 +1063,14 @@ Notes:
 // @SINGLE_MARKER
 export function mapPropObject<T extends object, K extends keyof T, Value>(
   valueMapper: (
-    value: T[K][number],
+    value: T[K] extends ReadonlyArray<infer ElementType> ? ElementType : never,
     data: T[K],
   ) => Value,
-	prop: K,
-): (data: T) => MergeTypes<Omit<T, K> & { [P in K]: Value[] }>;
-
+    prop: K,
+): (data: T) => T[K] extends ReadonlyArray<any>
+  ? MergeTypes<Omit<T, K> & { [P in K]: Value[] }>
+  : never;
+	
 /*
 Method: addPropToObjects