+++ index.d.ts
@@ -567,6 +567,19 @@ export function mapParallelAsync<T extends IterableContainer, U>(
 ): Promise<Mapped<T, U>>;
 
 /**
+ * It maps over a property of object that is a list.
+ */
+export function mapPropObject<T extends object, K extends keyof T, Value>(
+  valueMapper: (
+    value: T[K] extends ReadonlyArray<infer ElementType> ? ElementType : never,
+    data: T[K],
+  ) => Value,
+    prop: K,
+): (data: T) => T[K] extends ReadonlyArray<any>
+  ? MergeTypes<Omit<T, K> & { [P in K]: Value[] }>
+  : never;
+
+/**
  * Curried version of `String.prototype.match` which returns empty array, when there is no match.
  */
 export function match(regExpression: RegExp): (str: string) => string[];
@@ -1812,6 +1825,11 @@ export function rejectObject<T extends object>(
  * It replaces `strOrRegex` found in `str` with `replacer`.
  */
 export function replace(strOrRegex: RegExp | string, replacer: RegExp | string): (str: string) => string;
+
+/**
+ * Same as `R.replace` but it accepts array of string and regular expressions instead of a single value.
+ */
+export function replaceAll(patterns: (RegExp | string)[], replacer: string): (input: string) => string;
 
 /**
  * It returns a randomized copy of array.